# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AI Product Marketing Studio API
  version: 0.1.0
  description: Internal, workspace-scoped operations API. Provider secrets and raw provider payloads are never exposed.
servers:
  - url: /v1
security:
  - sessionCookie: []
paths:
...
  /operations/maintenance:
    put:
      operationId: setMaintenanceMode
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MaintenanceInput' } } } }
      responses:
        '200': { description: Maintenance mutation guard updated., content: { application/json: { schema: { $ref: '#/components/schemas/MaintenanceInput' } } } }
        '403': { $ref: '#/components/responses/Problem' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: studio_session
...
  responses:
    Problem:
      description: RFC 9457 problem details.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    MaintenanceInput:
      type: object
      additionalProperties: false
      required: [enabled, reason]
      properties:
        enabled: { type: boolean }
        reason: { type: string, maxLength: 500 }
...
    Problem:
      type: object
      additionalProperties: true
      required: [type, title, status, detail, requestId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        requestId:
          type: string
        errors:
          type: array
          items:
            type: object
            required: [field, code, message]
            properties:
              field:
                type: string
              code:
                type: string
              message:
                type: string